refactor(Detail): refactor pages/Detail folder
